% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{bar}
\alias{bar}
\title{Bar Plot Data}
\usage{
bar(D1a,D2a,D2b)
}
\description{
Visualizes the data
}
\examples{
data <- getGEO("GSE43452")
D2a <- extExp(data)[[2]] #gene expression
D2b <- extSample(data)[[2]] #sample annotations
D1a <- extGene(data)[[2]]
bar(D1a,D2a,D2b)

}
\author{
Nicholas Hutson
}
