% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{box}
\alias{box}
\title{Box Plot Data}
\usage{
box(D1a,D2a)
}
\description{
Visualizes the data
}
\examples{
data <- getGEO("GSE43452")
D2a <- extExp(data)[[2]]
D1a <- extGene(data)[[2]]
box(D1a,D2a)

}
\author{
Nicholas Hutson
}
