% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare Data}
\usage{
compare(D1a,D2a)
}
\description{
Compares the the genes with the top 10 variances and TP53 in a table
}
\examples{
data <- getGEO("GSE43452")
D2a <- extExp(data)[[2]]
D1a <- extGene(data)[[2]]
compare(D1a,D2a)

}
\author{
Nicholas Hutson
}
