% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extExp.R
\name{extExp}
\alias{extExp}
\title{Extract Expression Matrix}
\usage{
extExp(data, geneSymbol=NA, dName=NA)
}
\description{
Extracts the gene expression matrix
}
\examples{
extExp()

}
\author{
Nicholas Hutson
}
