% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extExpRSE.R
\name{extExpRSE}
\alias{extExpRSE}
\title{Extract Expression Matrix}
\usage{
extExpRSE(data, geneSymbol=NA, long = FALSE)
}
\arguments{
\item{data}{data imported using recount2}

\item{geneSymbol}{the gene "name" or "symbol"}

\item{long}{a boolean value entered for the data returned to be in a long format}
}
\description{
Extracts the gene expression matrix for RSE data
}
\examples{
download_study("SRP009615")
y <- file.path("SRP009615", 'rse_gene.Rdata')
load(y)
geneSymbol <- "TP53"
long <- TRUE
extExpRSE(rse_gene, geneSymbol, long)

}
\author{
Nicholas Hutson
}
