% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extGene.R
\name{extGene}
\alias{extGene}
\title{Extract Gene Annotation Matrix}
\usage{
extGene(data, geneSymbol = NA)
}
\arguments{
\item{data}{data imported using GEOquery}

\item{geneSymbol}{the gene "name" or "symbol"}
}
\description{
Extracts the gene annotation matrix
}
\examples{
data <- getGEO("GSE443452")
geneSymbol <- "TP53"
extGene(data, geneSymbol)

}
\author{
Nicholas Hutson
}
