% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extGeneGEO.R
\name{extGeneGEO}
\alias{extGeneGEO}
\title{Extract Gene Annotation Matrix}
\usage{
extGeneGEO(data, geneSymbol = NA)
}
\arguments{
\item{data}{data imported using GEOquery}

\item{geneSymbol}{the gene "name" or "symbol"}
}
\description{
Extracts the gene annotation matrix for GEO data
}
\examples{
data <- getGEO("GSE443452")
geneSymbol <- "TP53"
extGeneGEO(data, geneSymbol)

}
\author{
Nicholas Hutson
}
