% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extSampleGEO.R
\name{extSampleGEO}
\alias{extSampleGEO}
\title{Extract Sample Matrix}
\usage{
extSampleGEO(data,sampleName)
}
\arguments{
\item{data}{data imported using GEOquery}

\item{sampleName}{the sample name}
}
\description{
Extracts the sample annotation matrix for GEO data
}
\examples{
data <- getGEO("GSE443452")
sampleName <- "DBTRG untreated rep1"
extSampleGEO(data,sampleName)

}
\author{
Nicholas Hutson
}
