% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{hist}
\alias{hist}
\title{Histogram of Data}
\usage{
hist(D1a,D2a)
}
\description{
Compares the data in a histogram
}
\examples{
data <- getGEO("GSE43452")
D2a <- extExp(data)[[2]]
D1a <- extGene(data)[[2]]
hist(D1a,D2a)

}
\author{
Nicholas Hutson
}
