\name{print.GeneSummary_info}
\alias{print.GeneSummary_info}
\title{
Print the GeneSummary_info object
}
\description{
Print the GeneSummary_info object
}
\usage{
\method{print}{GeneSummary_info}(x, ...)
}
\arguments{

  \item{x}{A \code{GeneSummary_info} object.}
  \item{...}{Other arguments}

}
\value{
No value is returned.
}
\examples{
GeneSummary
}
