% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{Run Shiny app}
\usage{
runShiny()
}
\description{
Runs a Shiny app that allows the user to search the NCBI GEO database for data sets and creates data tables from them.
}
\examples{
runShiny()

}
\author{
Nicholas Hutson
}
