% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extExp.R
\name{extExp}
\alias{extExp}
\title{Extract Expression Matrix}
\usage{
extExp(data, geneSymbol=NA)
}
\arguments{
\item{data}{data imported using GEOquery}

\item{geneSymbol}{the gene "name" or "symbol"}
}
\description{
Extracts the gene expression matrix
}
\examples{
data <- getGEO("GSE443452")
geneSymbol <- "TP53"
extExp(data, geneSymbol)

}
\author{
Nicholas Hutson
}
